PluggableFileWatcher extension point
