+++ platform/platform-resources/src/META-INF/PlatformExtensionPoints.xml
@@ -244,5 +244,7 @@
     <extensionPoint name="runConfigurationBeforeRunProviderDelegate" interface="com.intellij.execution.impl.RunConfigurationBeforeRunProviderDelegate"/>
 
     <extensionPoint name="jbProtocolCommand" interface="com.intellij.openapi.application.JBProtocolCommand"/>
+
+    <extensionPoint name="vfs.local.pluggableFileWatcher" interface="com.intellij.openapi.vfs.impl.local.PluggableFileWatcher" />
   </extensionPoints>
 </idea-plugin>
