+++ platform/platform-resources/src/META-INF/PlatformExtensions.xml
@@ -400,5 +400,6 @@
 
     <ui.suitableFontProvider implementation="com.intellij.ui.SuitableFontProviderImpl"/>
     <jbProtocolCommand implementation="com.intellij.openapi.project.impl.JBProtocolOpenProjectCommand"/>
+    <vfs.local.pluggableFileWatcher implementation="com.intellij.openapi.vfs.impl.local.FileWatcherImpl" />
   </extensions>
 </idea-plugin>
